#41111 Split out webjobs-to-keep into separate classes
