+++ GenderPayGap.WebJob/Functions/FetchCompaniesHouseDataJob.cs
@@ -1,6 +1,8 @@
 ﻿using System;
 using System.Linq;
+using GenderPayGap.BusinessLogic.Services;
 using GenderPayGap.Core.Classes.Logger;
+using GenderPayGap.Core.Interfaces;
 using GenderPayGap.Database;
 using GenderPayGap.Extensions;
 using GenderPayGap.Extensions.AspNetCore;
@@ -8,8 +10,19 @@ using Microsoft.Azure.WebJobs;
 
 namespace GenderPayGap.WebJob
 {
-    public partial class Functions
+    public class FetchCompaniesHouseDataJob
     {
+        private readonly IDataRepository dataRepository;
+        private readonly UpdateFromCompaniesHouseService updateFromCompaniesHouseService;
+
+        public FetchCompaniesHouseDataJob(
+            IDataRepository dataRepository,
+            UpdateFromCompaniesHouseService updateFromCompaniesHouseService)
+        {
+            this.dataRepository = dataRepository;
+            this.updateFromCompaniesHouseService = updateFromCompaniesHouseService;
+        }
+
 
         [Singleton(Mode = SingletonMode.Listener)]
         public void FetchCompaniesHouseData([TimerTrigger("*/5 * * * *" /* evry 5 minutes */)] TimerInfo timer)
@@ -37,7 +50,7 @@ namespace GenderPayGap.WebJob
 
             for (var i = 0; i < maxNumCallCompaniesHouseApi; i++)
             {
-                long organisationId = _DataRepository.GetAll<Organisation>()
+                long organisationId = dataRepository.GetAll<Organisation>()
                     .Where(org => !org.OptedOutFromCompaniesHouseUpdate && org.CompanyNumber != null && org.CompanyNumber != "")
                     .OrderByDescending(org => org.LastCheckedAgainstCompaniesHouse == null)
                     .ThenBy(org => org.LastCheckedAgainstCompaniesHouse)
@@ -45,7 +58,7 @@ namespace GenderPayGap.WebJob
                     .FirstOrDefault();
 
                 CustomLogger.Information($"Start update companies house data organisation id: {organisationId}. Run id: {runId}");
-                _updateFromCompaniesHouseService.UpdateOrganisationDetails(organisationId);
+                updateFromCompaniesHouseService.UpdateOrganisationDetails(organisationId);
                 CustomLogger.Information($"End update companies house data organisation id: {organisationId}. Run id: {runId}");
             }
         }
